Move resolve_tilde and get_config_path into libi3
